+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/actions/JasyptAction.java
@@ -1,5 +1,6 @@
 package org.thermoweb.intellij.plugin.encrypt.actions;
 
+import java.util.Map;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -17,6 +18,7 @@ import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 
 import static org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog.ENCAPSULATE_FIELD_NAME;
+import static org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog.REMEMBER_PASSWORD;
 
 public class JasyptAction extends AnAction {
     private final Pattern pattern = Pattern.compile("ENC\\((.*)\\)");
@@ -63,8 +65,7 @@ public class JasyptAction extends AnAction {
     }
 
     protected void updateSettings() {
-        JasyptPluginSettings settings = project.getService(JasyptPluginSettings.class);
-        settings.algorithm = dialog.getValues().get(CipherInformationsDialog.ALGORITHM_FIELD_NAME);
-        settings.isEncapsulated = "true".equals(dialog.getValues().get(ENCAPSULATE_FIELD_NAME));
+        Map<String, String> values = dialog.getValues();
+        JasyptPluginSettings.updateSettings(project, values);
     }
 }